+++ CHANGELOG.md
@@ -6,6 +6,12 @@ This project adheres to [Semantic Versioning](http://semver.org/) and the [Keep 
 
 > Looking for the history of changes from upstream [elixirs/faker](https://github.com/elixirs/faker) (versions through `0.19.0-alpha.1`)? See [CHANGELOG-PRE-1.0.md](CHANGELOG-PRE-1.0.md).
 
+## Unreleased
+
+### Fixed
+
+- `Faker.Vehicle.En` - ensure generated VINs have the correct check digit [[@rubysolo](https://github.com/rubysolo)]
+
 ## [1.0.1](https://github.com/artkay/fakerer/compare/v1.0.0...v1.0.1) (2026-05-12)
 
 ### Fixed